Complete A-MoreOOP.
Classes are extracted.
