+++ text-ui-test/EXPECTED.TXT
@@ -17,13 +17,13 @@ ____________________________________________________________
 
 ____________________________________________________________
 Got it. I've added this task:
-  [D][ ] 333  (by: date1)
+  [D][ ] 333  (by: Dec-10-2012 12:10)
 Now you have 3 tasks in the list.
 ____________________________________________________________
 
 ____________________________________________________________
 Got it. I've added this task:
-  [E][ ] 444  (at: date2)
+  [E][ ] 444  (at: Oct-28-2030 13:12)
 Now you have 4 tasks in the list.
 ____________________________________________________________
 
@@ -31,8 +31,8 @@ ____________________________________________________________
 Here are the tasks in your list:
 1.[T][ ] 111
 2.[T][ ] 222
-3.[D][ ] 333  (by: date1)
-4.[E][ ] 444  (at: date2)
+3.[D][ ] 333  (by: Dec-10-2012 12:10)
+4.[E][ ] 444  (at: Oct-28-2030 13:12)
 ____________________________________________________________
 
 ____________________________________________________________
@@ -42,25 +42,25 @@ ____________________________________________________________
 
 ____________________________________________________________
 Nice! I've marked this task as done:
-  [D][X] 333  (by: date1)
+  [D][X] 333  (by: Dec-10-2012 12:10)
 ____________________________________________________________
 
 ____________________________________________________________
 Nice! I've marked this task as done:
-  [E][X] 444  (at: date2)
+  [E][X] 444  (at: Oct-28-2030 13:12)
 ____________________________________________________________
 
 ____________________________________________________________
 Here are the tasks in your list:
 1.[T][X] 111
 2.[T][ ] 222
-3.[D][X] 333  (by: date1)
-4.[E][X] 444  (at: date2)
+3.[D][X] 333  (by: Dec-10-2012 12:10)
+4.[E][X] 444  (at: Oct-28-2030 13:12)
 ____________________________________________________________
 
 ____________________________________________________________
 Noted. I've removed this task:
-  [E][X] 444  (at: date2)
+  [E][X] 444  (at: Oct-28-2030 13:12)
 Now you have 3 tasks in the list.
 ____________________________________________________________
 
@@ -68,7 +68,7 @@ ____________________________________________________________
 Here are the tasks in your list:
 1.[T][X] 111
 2.[T][ ] 222
-3.[D][X] 333  (by: date1)
+3.[D][X] 333  (by: Dec-10-2012 12:10)
 ____________________________________________________________
 
 exception.DukeException: ____________________________________________________________
