+++ text-ui-test/src/main/java/data/duke.txt
@@ -1,3 +1,3 @@
 [T][X] 111
 [T][ ] 222
-[D][X] 333  (by: date1)
+[D][X] 333  (by: Dec-10-2012 12:10)
